chore: generalise transaction state
